[Documentation] Trying TypeDoc to enhance documentation
One of the plus, provided by TypeDoc is that it supports function overload nicely.
As a consequence, tuple<T0>, tuple<T0,T1>... and other overloads appear under the same label.

At the same time, display is less pretty but it can be customized.
I just want to see if I can get something pretty with a bit of CSS magic and certainly tailwindcss.